pyproject.toml
[project]
name = "discord-github-bot"
version = "0.1.0"
description = "A Discord bot that creates GitHub issues from Discord messages"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "discord.py>=2.3.0",
    "PyGithub>=1.59.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]